3. The Hero Editor
You used the CLI to create a second HeroesComponent.
You displayed the HeroesComponent by adding it to the AppComponent
shell.
You applied the UppercasePipe to format the name.
You used two-way data binding with the ngModel directive.
You learned about the AppModule.
You imported the FormsModule in the AppModule so that Angular would
recognize and apply the ngModel directive.
You learned the importance of declaring components in the AppModule and
appreciated that the CLI declared it for you.